Move some classes to the test module.
